refactor: move utility functions to utils dir
