+++ src/utils/init-array.ts
